requirements.txt
pandas
numpy
jupyterlab
openai
sentence-transformers
faiss-cpu
